Updated game view to have the map view!!
Here we have the basic layout of the game view and map view! It is not super
cool looking yet but it's there. I decided to use a fixed size table so
that when everything moves around the table stays in place and looks OK.

I'll fix the way the boxes move around later as it is not mission critical!

Right now it generates a checker board patter but tomorrow it should be a
nice meadow.

Signed-off by Aidan Reidel areidel@pdx.edu