Small changes made to controller
Made some small changes to controller..didn't realize Wei already worked on this.  I just added a check for water.  I did uncomment the bottom section of the dead check.  This code was working before so there is no reason it should not work now.  Still needs to be tested but Wei's code worked when he commented out the bottom section.

Co-Authored-By: WEI <ironwall@users.noreply.github.com>