Added CharStatusView class
This is a basic implementation that displays character status (energy,
money, location) to text fields on the web page. I integrated it with
the mapview_test.html file. The front end html files are not in master
yet to integrate with, but it will be easy to do that once they arrive.

Signed-off-by: DanielBrewerPsu <brewer9@pdx.edu>