Added more Royal Diamonds code (good enough for Sprint 1)
More work related to those Royal Diamonds. I tested this using
mapview_test.html.

I added Royal Diamonds to the map that's built by Sample_Frupal.html.
So, to test this, start by sending your browser to maps/Sample_Frupal.
The diamonds are now 2 squares North of the starting position at 12, 14.
There's code in mapModel that stores the coordinates of the special
Royal Diamonds square when we're loading the map. By adding this code,
we're adding a way for maps to specify the Royal Diamonds' location;
before, there was a very crude way of specifying that location by
hard-coding it in the JS code instead of the map.

In the longer term, we should redo this after we implement the inventory
and items features. These Royal Diamonds are currently the one-and-only
special item in the game, but they ought to integrate with the other
items later (maybe in Sprint 2)

Signed-off-by: DanielBrewerPsu <brewer9@pdx.edu>