Initial Inventory View
It looks like there are a few more things to add to this, right now we are
simply displaying everything in the inventory list, but apparently the
inventory should include the price and quantity of each item. That will
require a bit more of a redesign of the inventory system.

Signed-off-by: David Post <post@pdx.edu>