Fixes to that Royal Diamonds disappear from map when you win
Fixes #62

I refactored this code related to Royal Diamonds. When this was added,
there was no inventory system, so Royal Diamonds needed to be handled as
a very special case: the only Item that the player picks up. Now, they
behave more like normal Binocs and other items. Because of this, the
MapModel and CharModel don't need special code to track their position.
This is slightly overkill for our needs, but not much. Another benefit
of this approach is that the Royal diamonds will be in the user's
inventory (once that display is working) on the Win screen.

Signed-off-by: DanielBrewerPsu <brewer9@pdx.edu>